// src/types/travel.ts
export type Language = 'en' | 'fr' | 'ar' | 'es' | 'zh';

export interface City {
  id: string;
  altitude: number;
  city_code: string;
  continent: string;
  country_code: string;
  have_airport: boolean;
  have_beach: boolean;
  have_historical_sites: boolean;
  have_low_crime_rate: boolean;
  have_mountains: boolean;
  have_nightlife: boolean;
  have_safe_drinking_water: boolean;
  image: string;
  image_lq: string;
  images: string[];
  is_atms_available: boolean;
  is_capital: boolean;
  l_ar_accessibility: string;
  l_ar_activities: string;
  l_ar_attractions: string;
  l_ar_climate: string;
  l_ar_emergency_numbers: string;
  l_ar_general_infos: string;
  l_ar_health: string;
  l_ar_languages: string;
  l_ar_name: string;
  l_ar_popular_dishes: string;
  l_ar_safety_infos: string;
  l_ar_transportation: string;
  l_ar_wikipedia: string;
  l_en_accessibility: string;
  l_en_activities: string;
  l_en_attractions: string;
  l_en_climate: string;
  l_en_emergency_numbers: string;
  l_en_general_infos: string;
  l_en_health: string;
  l_en_languages: string;
  l_en_name: string;
  l_en_popular_dishes: string;
  l_en_safety_infos: string;
  l_en_transportation: string;
  l_en_wikipedia: string;
  l_fr_accessibility: string;
  l_fr_activities: string;
  l_fr_attractions: string;
  l_fr_climate: string;
  l_fr_emergency_numbers: string;
  l_fr_general_infos: string;
  l_fr_health: string;
  l_fr_languages: string;
  l_fr_name: string;
  l_fr_popular_dishes: string;
  l_fr_safety_infos: string;
  l_fr_transportation: string;
  l_fr_wikipedia: string;
}

export interface Country {
  id: string;
  area: number;
  calling_code: string;
  continent: string;
  country_code: string;
  flag: string;
  flag_waiving: string;
  image: string;
  image_lq: string;
  l_ar_capital: string;
  l_ar_currency: string;
  l_ar_gastronomy: string;
  l_ar_general_infos: string;
  l_ar_geography: string;
  l_ar_history: string;
  l_ar_languages: string;
  l_ar_local_traditions: string;
  l_ar_name: string;
  l_ar_religions: string;
  l_ar_travel_advices: string;
  l_ar_unesco: string;
  l_ar_wikipedia: string;
  l_en_capital: string;
  l_en_currency: string;
  l_en_gastronomy: string;
  l_en_general_infos: string;
  l_en_geography: string;
  l_en_history: string;
  l_en_languages: string;
  l_en_local_traditions: string;
  l_en_name: string;
  l_en_religions: string;
  l_en_travel_advices: string;
  l_en_unesco: string;
  l_en_wikipedia: string;
  l_fr_capital: string;
  l_fr_currency: string;
  l_fr_gastronomy: string;
  l_fr_general_infos: string;
  l_fr_geography: string;
  l_fr_history: string;
}
// ...
export interface Landmark {
  id: string;
  city_code: string;
  country_code: string;
  image: string;
  images: string[];
  landmark_code: string;
  latitude: number;
  longitude: number;
  type: string;
  l_ar_description: string;
  l_ar_name: string;
  l_ar_wikipedia: string;
  l_en_address: string;
  l_en_description: string;
  l_en_name: string;
  l_en_types: string;
  l_en_wikipedia: string;
  l_es_description: string;
  l_es_name: string;
  l_es_wikipedia: string;
  l_fr_description: string;
  l_fr_name: string;
  l_fr_wikipedia: string;
  l_zh_description: string;
  l_zh_name: string;
  l_zh_wikipedia: string;
}